Support for darwin/arm Target
Hey all, was trying to use this provider on a darwin/arm machine, but it
looks like there is no release for it.

I added it to the `Makefile` here and was able to use it.

Additionally, I found that trying to build off `master` had an error
trying to reference a missing attribute of `MetricDataDataSource`, so
fixed that as well so the build would work.